+++ src/Controller/Sites.php
@@ -19,9 +19,9 @@ use Akeeba\Panopticon\Model\Reports;
 use Akeeba\Panopticon\Model\Site;
 use Akeeba\Panopticon\Model\Site as SiteModel;
 use Akeeba\Panopticon\Model\Task;
-use Akeeba\Panopticon\Task\EnqueueExtensionUpdateTrait;
-use Akeeba\Panopticon\Task\EnqueueJoomlaUpdateTrait;
 use Akeeba\Panopticon\Task\RefreshSiteInfo;
+use Akeeba\Panopticon\Task\Trait\EnqueueExtensionUpdateTrait;
+use Akeeba\Panopticon\Task\Trait\EnqueueJoomlaUpdateTrait;
 use Awf\Inflector\Inflector;
 use Awf\Mvc\DataController;
 use Awf\Registry\Registry;
