+++ src/Model/Site.php
@@ -20,8 +20,10 @@ use Akeeba\Panopticon\Exception\SiteConnection\SelfSignedSSL;
 use Akeeba\Panopticon\Exception\SiteConnection\SSLCertificateProblem;
 use Akeeba\Panopticon\Exception\SiteConnection\WebServicesInstallerNotEnabled;
 use Akeeba\Panopticon\Library\Task\Status;
-use Akeeba\Panopticon\Task\ApiRequestTrait;
+use Akeeba\Panopticon\Model\Trait\AdminToolsIntegrationTrait;
+use Akeeba\Panopticon\Model\Trait\AkeebaBackupIntegrationTrait;
 use Akeeba\Panopticon\Task\RefreshSiteInfo;
+use Akeeba\Panopticon\Task\Trait\ApiRequestTrait;
 use Awf\Container\Container;
 use Awf\Database\Query;
 use Awf\Date\Date;
