+++ src/Model/Trait/AkeebaBackupIntegrationTrait.php
@@ -5,7 +5,7 @@
  * @license   https://www.gnu.org/licenses/agpl-3.0.txt GNU Affero General Public License, version 3 or later
  */
 
-namespace Akeeba\Panopticon\Model;
+namespace Akeeba\Panopticon\Model\Trait;
 
 defined('AKEEBA') || die;
 
@@ -20,7 +20,7 @@ use Akeeba\Panopticon\Library\Task\Status;
 use Akeeba\Panopticon\Model\Exception\AkeebaBackupCannotConnectException;
 use Akeeba\Panopticon\Model\Exception\AkeebaBackupIsNotPro;
 use Akeeba\Panopticon\Model\Exception\AkeebaBackupNoInfoException;
-use Awf\Date\Date;
+use Akeeba\Panopticon\Model\Task;
 use Awf\Mvc\DataModel\Collection;
 use Awf\User\User;
 use Composer\CaBundle\CaBundle;
