+++ src/Task/FileScanner.php
@@ -14,7 +14,8 @@ use Akeeba\Panopticon\Library\Task\Attribute\AsTask;
 use Akeeba\Panopticon\Library\Task\Status;
 use Akeeba\Panopticon\Model\Reports;
 use Akeeba\Panopticon\Model\Site;
-use Awf\Mvc\Model;
+use Akeeba\Panopticon\Task\Trait\AdminToolsTrait;
+use Akeeba\Panopticon\Task\Trait\ApiRequestTrait;
 use Awf\Registry\Registry;
 use GuzzleHttp\RequestOptions;
