+++ src/Task/Trait/ApiRequestTrait.php
@@ -5,7 +5,7 @@
  * @license   https://www.gnu.org/licenses/agpl-3.0.txt GNU Affero General Public License, version 3 or later
  */
 
-namespace Akeeba\Panopticon\Task;
+namespace Akeeba\Panopticon\Task\Trait;
 
 
 use Akeeba\Panopticon\Model\Site;
