+++ src/Task/Trait/EnqueueJoomlaUpdateTrait.php
@@ -5,7 +5,7 @@
  * @license   https://www.gnu.org/licenses/agpl-3.0.txt GNU Affero General Public License, version 3 or later
  */
 
-namespace Akeeba\Panopticon\Task;
+namespace Akeeba\Panopticon\Task\Trait;
 
 defined('AKEEBA') || die;
 
@@ -13,8 +13,6 @@ use Akeeba\Panopticon\Container;
 use Akeeba\Panopticon\Library\Task\Status;
 use Akeeba\Panopticon\Model\Site;
 use Akeeba\Panopticon\Model\Task;
-use Awf\Date\Date;
-use Awf\Mvc\Model;
 use Awf\Registry\Registry;
 use Awf\User\User;
 use DateTimeZone;
